switch filesize to its own event handler

+++ src/videoobject.h
@@ -148,9 +148,7 @@ private:
     QStringList videoTrackList;
 
     QString currentFileName;
-    int currentFileSize;
-
-
+    qlonglong currentFileSize;
 };
 
 #endif // VIDEOOBJECT_H
